saas-bot-clean/requirements.txt
flask
flask-sqlalchemy
requests
google-genai
psutil
gunicorn
flask-limiter
python-dotenv
tenacity